Handle camera in Volume
Previoulsy when finding the start of the path to walk for a ray, the case where the camera was in the Volume was
not handled. This leads to increased processing time (walking voxels behind the vamera) and inaccurate rendering
at higher resolutions where intersections were recorded as being behind the camra - this may be indicative of a
probblem with integration too.

Now a test is performed to see if the origin is inside the voxel grid and if so, t is set to 0 and the start point
is set to the camera origin

Also put tighter checking around memory allocation, removing asserts and replacing with exceptions

+++ KinFu/TSDFVolume.hpp
@@ -207,8 +207,8 @@ namespace phd
          * @param z The depth coord (0-depth - 1)
          * @return te coordinate of the centre of the given voxel in world coords (m)
          */
-        inline size_t index( uint16_t x, uint16_t y, uint16_t z ) const {
-            size_t idx = z * m_xy_slice_size + y * m_x_size + x;
+        inline uint64_t index( uint16_t x, uint16_t y, uint16_t z ) const {
+            uint64_t idx = z * m_xy_slice_size + y * m_x_size + x;
             
             return idx;
         }
